backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
pyrebase4>=4.8.0
web3>=6.15.0
opencv-python>=4.9.0
numpy>=1.26.0
cloudinary>=1.36.0
